refactor: mypage 페키지 분리하기
