+++ src/main/java/com/soolsul/soolsulserver/auth/presentation/AuthController.java
@@ -1,15 +1,11 @@
 package com.soolsul.soolsulserver.auth.presentation;
 
-import com.soolsul.soolsulserver.auth.CustomUser;
 import com.soolsul.soolsulserver.auth.business.CustomUserDetailsService;
 import com.soolsul.soolsulserver.auth.presentation.dto.RegisterRequest;
-import com.soolsul.soolsulserver.auth.repository.dto.UserLookUpResponse;
 import com.soolsul.soolsulserver.common.response.BaseResponse;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.security.core.Authentication;
-import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -18,7 +14,6 @@ import org.springframework.web.bind.annotation.RestController;
 import javax.validation.Valid;
 
 import static com.soolsul.soolsulserver.common.response.ResponseCodeAndMessages.USER_CREATE_SUCCESS;
-import static com.soolsul.soolsulserver.common.response.ResponseCodeAndMessages.USER_LOOK_UP_SUCCESS;
 
 @RequiredArgsConstructor
 @RestController
@@ -31,12 +26,5 @@ public class AuthController {
     public ResponseEntity<BaseResponse<Void>> registerUser(@Valid @RequestBody RegisterRequest registerRequest) {
         userDetailsService.register(registerRequest);
         return new ResponseEntity<>(new BaseResponse<>(USER_CREATE_SUCCESS, null), HttpStatus.CREATED);
-    }
-
-    @GetMapping("/me")
-    public ResponseEntity<BaseResponse<UserLookUpResponse>> searchUser(Authentication authentication) {
-        CustomUser user = (CustomUser) authentication.getPrincipal();
-        UserLookUpResponse userLookUpResponse = userDetailsService.findUserWithDetailInfo(user.getId());
-        return ResponseEntity.ok(new BaseResponse<>(USER_LOOK_UP_SUCCESS, userLookUpResponse));
     }
 }
